move patricia package under seg
