+++ erigon-lib/seg/patricia/patricia.go
@@ -21,9 +21,10 @@ import (
 	"math/bits"
 	"strings"
 
+	"golang.org/x/exp/slices"
+
 	"github.com/ledgerwatch/erigon-lib/common/cmp"
 	"github.com/ledgerwatch/erigon-lib/sais"
-	"golang.org/x/exp/slices"
 )
 
 // Implementation of paticia tree for efficient search of substrings from a dictionary in a given string
